An R6 Class for estimating an probability matrix from an R x C ecological inference problem.

This class implements an EM algorithm using different methods to approximate the E-step, such as "Multinomial", "Hit and Run", "MVN CDF", "MVN PDF", and "Exact".

# S3 method for class 'eim'
summary(object)

# S3 method for class 'eim'
predict(object, ...)

# S3 method for class 'eim'
as.matrix(object)

sd.eim(object, ...)

Note

Precomputing can eventually accelerate the computation time of the Expectation-Maximization algorithm.

The $compute, $bootstrap, $print and $summary methods can also be called as an S3 method with predict(), sd(), print() and summary().

References

Thraves, C. and Ubilla, P.: 'Fast Ecological Inference Algorithm for the R×C Case'

Author

Charles Thraves, Pablo Ubilla

Public fields

X (matrix)	An (b x c) matrix with the observed results of the candidate votes (c)

for each ballot box (b). Provided manually or loaded from JSON.

W (matrix)

An (b x g) matrix with the observed results of the demographic group

votes (g) for each ballot box (b). Provided manually or loaded from JSON.

method (character)	A string indicating the EM method. One of: "Multinomial", "Hit and Run",

"MVN CDF", "MVN PDF", or "Exact".

probability (matrix)

An (g x c) matrix that will store the final estimated probabilities of a given

group (g) voting for a candidate (c).

logLikelihood (numeric)	A numeric vector that will store the log-likelihood values over all iterations

of the Expectation-Maximization algorithm.

total_iterations (integer(1))

An integer indicating the total number of iterations performed by the

Expectation-Maximization algorithm.

total_time (numeric(1))	The time taken by the EM algorithm.

finish_state (character)

The reason for the algorithm's termination. It can be either

"Maximum iterations", "Log-likelihood decrease", "Convergence",

Active bindings

samples (integer(1))	Active variable to show the Hit and Run samples if and

only if the method is "Hit and Run".

step_size (integer(1))

Active variable to show the Hit and Run step size if and

only if the method is "Hit and Run".

multivariate_method (character)	Active variable to show the method used to estimate

the Multivariate Normal CDF, if and only if self$method is

"MVN CDF".

multivariate_error (numeric(1))

Active variable to show the error threshold for the Monte Carlo

simulation of the Multivariate Normal CDF.

multivariate_iterations (numeric(1))	Active variable to show the number of iterations for the Monte

Carlo simulation of the Multivariate Normal CDF.

sd (matrix)

Active variable to show an estimate of the probability standard

Methods

Public methods

eim$new()

eim$precompute()

eim$compute()

eim$bootstrap()

eim$print()

eim$summary()

eim$save_results()

eim$clone()

Method new()

Creates the object by defining the X and W matrix attributes.

Usage

eim$new(X = NULL, W = NULL, json_path = NULL)

Arguments

X (matrix)	A matrix (c x b) of observed candidate votes per ballot box (b)

(optional; required if json_path is NULL).

W (matrix)

A matrix (b x g) of observed demographic group votes per

ballot box (b) (optional; required if json_path is NULL).

json_path (character)	A string containing a path to a JSON file with "X" and "W"

matrices. (optional; required if X or W are NULL)

Returns

An initialized eim object.

Examples

# Example 1: Create a eim object from a matrix
model <- eim$new(X = matrix(1:9, 3, 3), W = matrix(1:9, 3, 3))

# Example 2: Create a eim object from a JSON file
\dontrun{
model2 <- eim$new(json_path = "a/file/path/to/a/file.json")
}

Method precompute()

Calculates the EM-independent variables for the Hit and Run and Exact methods.

Usage

eim$precompute(method, ...)

Arguments

method (character)	The method for precomputing. Options:

"Hit and Run" or "Exact".

...

Additional arguments required by specific methods:

"Hit and Run" Method:

step_size (integer(1))

The step size (M) for the Hit and Run algorithm.

Must be a positive integer.

samples (integer(1))	The number of samples (S) to generate. Must be an

integer.

Returns

The modified eim object (for method chaining). Updates are made on the C internal memory.

Examples

# Example 1: Precompute the Hit and Run method
simulations <- simulate_elections(num_ballots = 20,
        num_candidates = 5,
        num_groups = 3,
        ballot_voters = rep(100, 20))

model <- eim$new(simulations$X, simulations$W)

model$precompute("Hit and Run",
          step_size = 1000,
          samples = 5000)
# Changes are made to the C internals API

# Example 2: Precompute the Exact method
model$precompute("Exact")

Method compute()

Executes the Expectation-Maximization (EM) algorithm based on the selected method. Additional parameters may be required depending on the method.

Usage

eim$compute(
  method = "Multinomial",
  probability_method = "Group proportional",
  iterations = 1000,
  maximum_time = 1440,
  stopping_threshold = 0.001,
  verbose = FALSE,
  ...
)

Arguments

method (character)	The method for estimating the

Expectation-Maximization (EM) algorithm.

Options: "Multinomial", "MVN CDF",

"MVN PDF", "Hit and Run",

and "Exact"

(default: "Multinomial").

probability_method (character)

The method for obtaining the initial

probability. Options: "Group proportional",

"Proportional", or "Uniform".

(default: "Group proportional").

iterations (integer(1))	The maximum number of iterations to

perform on the EM algorithm.

(default: 1000).

maximum_time (integer(1))

The maximum time (in minutes) that the

algorithm will iterate. (default: 1440).

stopping_threshold (numeric(1))	The minimum difference between successive

probabilities for stopping the iterations.

(default: 0.001).

verbose (boolean(1))

A boolean indicating whether to print

informative messages while iterating.

(default: FALSE).

...	Additional arguments required by specific

methods:

"Hit and Run" Method:

step_size (integer(1)):

The step size (M) for the Hit and Run

algorithm.

samples (integer(1)):	The number of samples (S) to generate.

"MVN CDF" Method:

multivariate_method (character):

The integration method. Can be chosen

between "Genz" and "Genz2". Default is

"Genz2".

multivariate_error (numeric(1)):	The integration error threshold. Default is

1e-6.

multivariate_iterations (integer(1)):

The number of Monte Carlo iterations.

Returns

The modified eim object (for method chaining).

Examples

# Example 1: Compute the Expectation-Maximization with default values

simulations <- simulate_elections(num_ballots = 20,
				num_candidates = 5,
				num_groups = 3,
				ballot_voters = rep(100, 20))

model <- eim$new(simulations$X, simulations$W)
model$compute() # Retrieves the object with updated attributes

# Example 2: Compute the Expectation-Maximization for the Hit and Run method
\donttest{
	model$compute(method = "Hit and Run",
					step_size = 3000,
					samples = 1000)
	}

# Example 3: Omit arguments to the Hit and Run method
\dontrun{
	model$compute(method = "Hit and Run",
					step_size = 3000)
	# Error; must supply the samples parameter as well
	}

# Example 4: Run the MVN CDF with default values
model$compute(method = "MVN CDF")

# Example 5: Run an Exact estimation with user-defined parameters
model$compute(method = "Exact",
				probability_method = "Uniform",
 			iterations = 5,
				stopping_threshold = 1e-3)
# Verbose was omitted

Method bootstrap()

Runs a bootstrap algorithm to estimate the standard deviation of probabilities.

Usage

eim$bootstrap(
  bootstrap_iterations,
  ballot_boxes,
  method = "Multinomial",
  probability_method = "Group proportional",
  iterations = 1000,
  stopping_threshold = 0.001,
  verbose = FALSE,
  ...
)

Arguments

bootstrap_iterations (integer(1))	The number of EM computations.

ballot_boxes (integer(1))

The number of ballot boxes to use

as a sample. It must be

strictly less than the total

number of ballot boxes.

Arguments for calling the $compute()	method:

method (character)

The method for estimating the

Expectation-Maximization (EM) algorithm.

Options: "Multinomial", "MVN CDF",

"MVN PDF", "Hit and Run",

and "Exact"

(default: "Multinomial").

probability_method (character)	The method for obtaining the initial

probability. Options: "Group proportional",

"Proportional", or "Uniform".

(default: "Group proportional").

iterations (integer(1))

The maximum number of iterations to

perform on the EM algorithm.

(default: 1000).

maximum_time (integer(1))	The maximum time (in minutes) that the

algorithm will iterate. (default: 1440).

stopping_threshold (numeric(1))

The minimum difference between successive

probabilities for stopping the iterations.

(default: 0.001).

verbose (boolean(1))	A boolean indicating whether to print

informative messages while iterating.

(default: FALSE).

...

Additional arguments required by specific

methods:

"Hit and Run" Method:

step_size (integer(1)):	The step size (M) for the Hit and Run

algorithm.

samples (integer(1)):

The number of samples (S) to generate.

"MVN CDF" Method:

multivariate_method (character):	The integration method. Can be chosen

between "Genz" and "Genz2". Default is

"Genz2".

multivariate_error (numeric(1)):

The integration error threshold. Default is

1e-6.

multivariate_iterations (integer(1)):	The number of Monte Carlo iterations.

Default is 5000.

Returns

The modified eim object, under the field $sd.

Examples

simulations <- simulate_elections(num_ballots = 20,
        num_candidates = 5,
        num_groups = 3,
        ballot_voters = rep(100, 20))

model <- eim$new(simulations$X, simulations$W)

model$bootstrap(30, 10)

model$sd # An estimate of the standard deviation of the probabilities.

Method print()

Depending on the state of the algorithm (computed or not), it prints a message with its most relevant parameters.

Usage

eim$print()

Returns

The object itself (for method chaining).

Examples

simulations <- simulate_elections(num_ballots = 20,
        num_candidates = 5,
        num_groups = 3,
        ballot_voters = rep(100, 20))

model <- eim$new(simulations$X, simulations$W)
print(model) # Prints the X and W matrices.

model$compute()
print(model) # Prints the X and W matrices along with the EM results.

Method summary()

Shows, in the form of a list, a selection of the most important attributes. It retrieves the method, number of candidates, ballots, and groups, as well as the principal results of the EM algorithm.

Usage

eim$summary()

Returns

(list) A list with the method, candidates, ballots, groups, probabilities, log-likelihood, finish_state and standard deviation.

Examples

simulations <- simulate_elections(num_ballots = 5,
        num_candidates = 3,
        num_groups = 2,
        ballot_voters = rep(100, 5))

model <- eim$new(simulations$X, simulations$W)
a_list <- model$summary()
a_list$method # Not computed yet
a_list$groups # 2
a_list$ballots # 5
names(a_list)
# "candidates" "groups" "ballots" "method" "probabilities"
# "logLikelihood" "finish_state" "sd"

Method save_results()

Saves the current eim object to a specified file. The results can be saved in:

RDS (Binary format): Preserves object structure for future use in R.

JSON: Saves model data in a human-readable format.

CSV: Saves the probability matrix in a tabular format.

Usage

eim$save_results(filename)

Arguments

filename (character)	The file name where the results should be saved,

including its extension. The file

extension determines the format:

*.rds

Saves as RDS (default, binary format).

*.json	Saves as JSON (readable and shareable).

*.csv

Returns

The modified eim object (for method chaining).

Examples

simulations <- simulate_elections(num_ballots = 20,
        num_candidates = 5,
        num_groups = 3,
        ballot_voters = rep(100, 20))

model <- eim$new(simulations$X, simulations$W)
model$compute()
model$save_results("results.rds")  # Save as RDS
model$save_results("results.json") # Save as JSON
model$save_results("results.csv")  # Save the final probability matrix as CSV

Method clone()

The objects of this class can be cloned using this method.

Usage

eim$clone(deep = FALSE)

Arguments

deep (bool(1))	Whether to make a deep clone.

Examples

## ------------------------------------------------
## Method `eim$new`
## ------------------------------------------------

# Example 1: Create a eim object from a matrix
model <- eim$new(X = matrix(1:9, 3, 3), W = matrix(1:9, 3, 3))

# Example 2: Create a eim object from a JSON file
if (FALSE) { # \dontrun{
model2 <- eim$new(json_path = "a/file/path/to/a/file.json")
} # }

## ------------------------------------------------
## Method `eim$precompute`
## ------------------------------------------------

# Example 1: Precompute the Hit and Run method
simulations <- simulate_elections(num_ballots = 20,
        num_candidates = 5,
        num_groups = 3,
        ballot_voters = rep(100, 20))

model <- eim$new(simulations$X, simulations$W)

model$precompute("Hit and Run",
          step_size = 1000,
          samples = 5000)
#> Precomputing the Hit and Run method
# Changes are made to the C internals API

# Example 2: Precompute the Exact method
model$precompute("Exact")
#> Precomputing the Exact method

## ------------------------------------------------
## Method `eim$compute`
## ------------------------------------------------

# Example 1: Compute the Expectation-Maximization 
# with default values

simulations <- simulate_elections(num_ballots = 20,
        num_candidates = 5,
        num_groups = 3,
        ballot_voters = rep(100, 20))

model <- eim$new(simulations$X, simulations$W)
model$compute() # Retrieves the object with updated 
# attributes

# Example 2: Compute the Expectation-Maximization 
# for the Hit and Run method
# \donttest{
  model$compute(method = "Hit and Run",
          step_size = 3000,
          samples = 1000)
  # }

# Example 3: Omit arguments to the Hit and Run 
# method
if (FALSE) { # \dontrun{
  model$compute(method = "Hit and Run",
          step_size = 3000)
  # Error; must supply the samples parameter as well
  } # }

# Example 4: Run the MVN CDF with default values
model$compute(method = "MVN CDF")

# Example 5: Run an Exact estimation with 
# user-defined parameters
model$compute(method = "Exact",
        probability_method = "Uniform",
       iterations = 5,
        stopping_threshold = 1e-3)
# Verbose was omitted

## ------------------------------------------------
## Method `eim$bootstrap`
## ------------------------------------------------

simulations <- simulate_elections(num_ballots = 20,
        num_candidates = 5,
        num_groups = 3,
        ballot_voters = rep(100, 20))

model <- eim$new(X = simulations$X, W = simulations$W)

model$bootstrap(30, 10)

model$sd # An estimate of the probabilities' standard 
#>            [,1]       [,2]       [,3]        [,4]        [,5]
#> [1,] 0.04800395 0.05681623 0.01749571 0.017243044 0.003316897
#> [2,] 0.02954469 0.03514758 0.03982698 0.020104261 0.017494917
#> [3,] 0.02139005 0.03856379 0.01917266 0.007544839 0.024901741
# deviation.

## ------------------------------------------------
## Method `eim$print`
## ------------------------------------------------

simulations <- simulate_elections(num_ballots = 20,
        num_candidates = 5,
        num_groups = 3,
        ballot_voters = rep(100, 20))

model <- eim$new(simulations$X, simulations$W)
print(model) # Prints the X and W matrices.
#> eim ecological inference model
#> Candidate matrix (X) [b x c]:
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]   32   23   19   15   11
#> [2,]   32   17   32   10    9
#> [3,]   30   17   26    9   18
#> [4,]   35   17   23   11   14
#> [5,]   29   17   28   11   15
#> ...
#> Group matrix (W) [b x g]:
#>      [,1] [,2] [,3]
#> [1,]   69   16   15
#> [2,]   70   15   15
#> [3,]   74   13   13
#> [4,]   65   18   17
#> [5,]   60   18   22
#> ...

model$compute()
print(model) # Prints the X and W matrices along with 
#> eim ecological inference model
#> Candidate matrix (X) [b x c]:
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]   32   23   19   15   11
#> [2,]   32   17   32   10    9
#> [3,]   30   17   26    9   18
#> [4,]   35   17   23   11   14
#> [5,]   29   17   28   11   15
#> ...
#> Group matrix (W) [b x g]:
#>      [,1] [,2] [,3]
#> [1,]   69   16   15
#> [2,]   70   15   15
#> [3,]   74   13   13
#> [4,]   65   18   17
#> [5,]   60   18   22
#> ...
#> Estimated probability [g x c]:
#>            [,1]      [,2]       [,3]       [,4]       [,5]
#> [1,] 0.45769715 0.1022744 0.24631703 0.08282288 0.11088857
#> [2,] 0.04050710 0.6067338 0.02557169 0.26854714 0.05864029
#> [3,] 0.02915878 0.1348679 0.52309578 0.05962236 0.25325517
#> Method:	 Multinomial 
#> Total Iterations:	 22 
#> Total Time (s):	 6.5e-05 
#> Log-likelihood:	 -22.76561
# the EM results.

## ------------------------------------------------
## Method `eim$summary`
## ------------------------------------------------

simulations <- simulate_elections(num_ballots = 5,
        num_candidates = 3,
        num_groups = 2,
        ballot_voters = rep(100, 5))

model <- eim$new(simulations$X, simulations$W)
a_list <- model$summary()
a_list$method # Not computed yet
#> [1] "Not computed yet"
a_list$groups # 2
#> [1] 2
a_list$ballots # 5
#> [1] 3
names(a_list)
#> [1] "candidates"    "groups"        "ballots"       "method"       
#> [5] "probabilities" "logLikelihood" "finish_state"  "sd"           
# "candidates" "groups" "ballots" "method" "probabilities" 
# "logLikelihood" "finish_state" "sd"

## ------------------------------------------------
## Method `eim$save_results`
## ------------------------------------------------

simulations <- simulate_elections(num_ballots = 20,
        num_candidates = 5,
        num_groups = 3,
        ballot_voters = rep(100, 20))

model <- eim$new(simulations$X, simulations$W)
model$compute()
model$save_results("results.rds")  # Save as RDS
#> Results saved as RDS: results.rds
model$save_results("results.json") # Save as JSON
#> Results saved as JSON: results.json
model$save_results("results.csv")  # Save the final 
#> Probability matrix saved as CSV: results.csv
# probability matrix as CSV