Randomly create a voting instance by defining an interval

Given a range of possible observed outcomes (such as the number of ballot boxes, candidates, etc.), it creates a completely random voting instance, simulating the unobserved results as well.

random_samples(
  ballots_range,
  candidates_range,
  group_range,
  voting_range,
  seed = NULL
)

Arguments

ballots_range

(integer) A vector of length 2 specifying the lower and upper bounds for the number of ballot boxes.

candidates_range

(integer) A vector of length 2 specifying the lower and upper bounds for the number of candidates to draw.

group_range

(integer) A vector of length 2 specifying the lower and upper bounds for the number of demographic groups to draw.

voting_range

(integer) A vector of length 2 specifying the lower and upper bounds for the number of votes per ballot box.

seed

(numeric(1)) Optional. If provided, it overrides the current global seed.

Value

A list with components:

X

A matrix (b x c) containing candidate votes per ballot box.

W

A matrix (b x g) containing demographic votes per ballot box.

real_p

A matrix (g x c) containing the estimated (unobserved) probabilities that a demographic group votes for a given candidate.

ballots

The number of ballot boxes that were drawn.

candidates

The number of candidates that were drawn.

groups

The number of demographic groups that were drawn.

total_votes

A vector containing the total number of votes per ballot box.

See also

Examples

bal_range <- c(30, 50)
can_range <- c(2, 4)
group_range <- c(2, 6)
voting_range <- c(50, 100)
results <- random_samples(bal_range, can_range, group_range, voting_range)

# X matrix
results$X # A randomly generated matrix of dimension (b x c)
#>       [,1] [,2]
#>  [1,]   41   16
#>  [2,]   69   17
#>  [3,]   57   15
#>  [4,]   76   21
#>  [5,]   74   17
#>  [6,]   75   23
#>  [7,]   43   10
#>  [8,]   54   21
#>  [9,]   67   18
#> [10,]   34   18
#> [11,]   63   15
#> [12,]   53   33
#> [13,]   54   15
#> [14,]   51   16
#> [15,]   44    8
#> [16,]   53   20
#> [17,]   21   32
#> [18,]   30   36
#> [19,]   23   36
#> [20,]   34   54
#> [21,]   31   49
#> [22,]   32   35
#> [23,]   21   31
#> [24,]   28   36
#> [25,]   18   34
#> [26,]   38   40
#> [27,]   28   38
#> [28,]   35   46
#> [29,]   39   59
#> [30,]   33   47
#> [31,]   21   33
#> [32,]   25   44
#> [33,]   33   58
#> [34,]   38   31
#> [35,]   43   51
#> [36,]   45   40
#> [37,]   49   31
#> [38,]   51   43
#> [39,]   34   38
#> [40,]   41   20
#> [41,]   30   22
#> [42,]   55   41
#> [43,]   54   33
#> [44,]   44   28
#> [45,]   27   28
#> [46,]   35   42
#> [47,]   59   40
#> [48,]   25   36
#> [49,]   24   27
ncol(results$X <= can_range[2]) # Always TRUE
#> [1] 2
ncol(results$X >= can_range[1]) # Always TRUE
#> [1] 2
nrow(results$X <= bal_range[2]) # Always TRUE
#> [1] 49
nrow(results$X >= bal_range[1]) # Always TRUE
#> [1] 49

# W matrix
results$W # A randomly generated matrix of dimension (b x g)
#>       [,1] [,2] [,3]
#>  [1,]   36   10   11
#>  [2,]   61   11   14
#>  [3,]   46   12   14
#>  [4,]   67   13   17
#>  [5,]   65   10   16
#>  [6,]   69   13   16
#>  [7,]   37    9    7
#>  [8,]   44   13   18
#>  [9,]   56   11   18
#> [10,]   33   13    6
#> [11,]   54    9   15
#> [12,]   50   21   15
#> [13,]   42   17   10
#> [14,]   40   10   17
#> [15,]   38    7    7
#> [16,]   47   14   12
#> [17,]   11   36    6
#> [18,]   10   48    8
#> [19,]   13   34   12
#> [20,]   14   65    9
#> [21,]   12   56   12
#> [22,]   15   41   11
#> [23,]    9   31   12
#> [24,]   17   36   11
#> [25,]    8   37    7
#> [26,]    8   54   16
#> [27,]    9   44   13
#> [28,]   12   53   16
#> [29,]   16   69   13
#> [30,]   17   49   14
#> [31,]    9   36    9
#> [32,]    8   47   14
#> [33,]   17   58   16
#> [34,]   14   15   40
#> [35,]   18   14   62
#> [36,]   13   13   59
#> [37,]    9   14   57
#> [38,]   15   16   63
#> [39,]    8   11   53
#> [40,]   12    4   45
#> [41,]   11   10   31
#> [42,]   20   18   58
#> [43,]   17   13   57
#> [44,]   13   13   46
#> [45,]    6   13   36
#> [46,]   17   13   47
#> [47,]   15   13   71
#> [48,]   11   17   33
#> [49,]    9   14   28
ncol(results$W <= group_range[2]) # Always TRUE
#> [1] 3
ncol(results$W >= group_range[1]) # Always TRUE
#> [1] 3
nrow(results$W <= bal_range[2]) # Always TRUE
#> [1] 49
nrow(results$W >= bal_range[1]) # Always TRUE
#> [1] 49

# Probability matrix
results$real_p # A matrix (g x c) that summarizes the unobserved outcomes
#>           [,1]       [,2]
#> [1,] 0.9366869 0.06331312
#> [2,] 0.2112752 0.78872477
#> [3,] 0.5667240 0.43327601
ncol(results$real_p) == ncol(results$X) # Always TRUE
#> [1] TRUE
nrow(results$real_p) == ncol(results$W) # Always TRUE
#> [1] TRUE