Update an existing eim model with a new EM algorithm computation

This function updates an object by performing a new Expectation-Maximization computation with different parameters.

# S3 method for class 'eim'
update(object, ...)

Arguments

...

New parameters to be passed to compute().

object

An eim object.

Value

The updated object with the changes applied.

Examples

simulations <- simulate_elections(
  num_ballots = 20,
  num_candidates = 5,
  num_groups = 3,
  ballot_voters = rep(100, 20)
)

model <- eim$new(simulations$X, simulations$W)

predict(model)
#>           [,1]       [,2]       [,3]      [,4]       [,5]
#> [1,] 0.1846717 0.15046501 0.23239874 0.3430691 0.08939542
#> [2,] 0.7623260 0.05402071 0.01371131 0.1391343 0.03080769
#> [3,] 0.1056694 0.29643583 0.12883015 0.2028004 0.26626420

update(model, "MVN PDF")
#> eim ecological inference model
#> Candidate matrix (X) [b x c]:
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]   28   17   18   29    8
#> [2,]   35   15   19   21   10
#> [3,]   31   16   13   29   11
#> [4,]   22   17   17   29   15
#> [5,]   22   14   21   30   13
#> ...
#> Group matrix (W) [b x g]:
#>      [,1] [,2] [,3]
#> [1,]   65   17   18
#> [2,]   60   24   16
#> [3,]   62   23   15
#> [4,]   70   10   20
#> [5,]   72    8   20
#> ...
#> Estimated probability [g x c]:
#>           [,1]       [,2]       [,3]      [,4]       [,5]
#> [1,] 0.1893925 0.14857844 0.23032944 0.3434990 0.08820060
#> [2,] 0.7481907 0.05855448 0.02353965 0.1377368 0.03197841
#> [3,] 0.1165076 0.29238131 0.12248985 0.2068617 0.26175953
#> Method:	 MVN PDF 
#> Total Iterations:	 10 
#> Total Time (s):	 0.117297 
#> Log-likelihood:	 -25.84529

model$method # Returns 'MVN PDF'
#> [1] "MVN PDF"